+++ diode-core/shared/src/main/scala-2.12/diode/EffectSetExecutionOps.scala
@@ -1,0 +1,9 @@
+package diode
+
+import scala.collection.generic.CanBuildFrom
+import scala.concurrent.Future
+
+trait EffectSetExecutionOps { self: EffectSet =>
+  private[diode] def executeWith[A](f: Effect => Future[A]): Future[Set[A]] =
+    Future.traverse(tail + head)(f(_))(implicitly[CanBuildFrom[Set[Effect], A, Set[A]]], ec)
+}
