+++ diode-core/shared/src/main/scala-2.13/diode/EffectSetExecutionOps.scala
@@ -1,0 +1,9 @@
+package diode
+
+import scala.concurrent.Future
+import scala.collection.BuildFrom
+
+trait EffectSetExecutionOps { self: EffectSet =>
+  private[diode] def executeWith[A](f: Effect => Future[A]): Future[Set[A]] =
+    Future.traverse(tail + head)(f(_))(implicitly[BuildFrom[Set[Effect], A, Set[A]]], ec)
+}
