+++ diode-core/shared/src/main/scala/diode/Effect.scala
@@ -138,6 +138,35 @@ class EffectSeq(head: Effect, tail: Seq[Effect], ec: ExecutionContext) extends E
     new EffectSeq(head.flatMap(g), tail.map(_.flatMap(g)), ec)
 }
 
+/**
+  * Wraps multiple `Effects` to be executed later. Effects are executed in parallel without any ordering.
+  *
+  * @param head First effect to be run.
+  * @param tail Rest of the effects.
+  */
+class EffectSet(val head: Effect, val tail: Set[Effect], ec: ExecutionContext)
+    extends EffectBase(ec)
+    with EffectSetExecutionOps {
+
+  override def run(dispatch: Any => Unit) =
+    executeWith(_.run(dispatch)).map(_ => ())(ec)
+
+  override def +(that: Effect) =
+    new EffectSet(head, tail + that, ec)
+
+  override def size =
+    head.size + tail.foldLeft(0)((acc, e) => acc + e.size)
+
+  override def toFuture =
+    executeWith(_.toFuture)
+
+  override def map[B: ActionType](g: Any => B) =
+    new EffectSet(head.map(g), tail.map(_.map(g)), ec)
+
+  override def flatMap[B: ActionType](g: Any => Future[B]) =
+    new EffectSet(head.flatMap(g), tail.map(_.flatMap(g)), ec)
+}
+
 object Effect {
   type EffectF[A] = () => Future[A]
